introduce WIth{Model, Entity}Trait traits
