+++ src/View/WithEntityTrait.php
@@ -1,0 +1,47 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Atk4\Ui\View;
+
+use Atk4\Data\Model;
+use Atk4\Ui\Exception;
+
+trait WithEntityTrait
+{
+    public ?Model $entity = null;
+
+    #[\Override]
+    public function &__get(string $name)
+    {
+        // TODO remove in atk4/ui 6.0
+        if ($name === 'model' && !(new \ReflectionProperty(self::class, 'model'))->isInitialized($this) && $this->entity !== null) {
+            throw new Exception('Use View::$entity property instead for entity access');
+        }
+
+        return parent::__get($name);
+    }
+
+    /**
+     * Associate this view with an entity. Do not place any logic in this method, instead take it to renderView().
+     *
+     * Do not try to create your own "Model" implementation, instead you must be looking for
+     * your own "Persistence" implementation.
+     *
+     * @phpstan-assert !null $this->entity
+     */
+    public function setEntity(Model $entity): void
+    {
+        $entity->assertIsEntity();
+
+        if ($this->entity !== null) {
+            if ($this->entity === $entity) {
+                return;
+            }
+
+            throw new Exception('Different entity is already set');
+        }
+
+        $this->entity = $entity;
+    }
+}
